core/requirements.txt
openai
python-dotenv
rich
duckduckgo-search
aiohttp
fastapi
uvicorn
google-genai
pydantic
requests